API

API

BED.Reader — Type.

BED.Reader(input::IO; index=nothing)
BED.Reader(input::AbstractString; index=:auto)

Create a data reader of the BED file format.

The first argument specifies the data source. When it is a filepath that ends with .bgz, it is considered to be block compression file format (BGZF) and the function will try to find a tabix index file (<filename>.tbi) and read it if any. See http://www.htslib.org/doc/tabix.html for bgzip and tabix tools.

Arguments

input: data source

index: path to a tabix file

source

BED.Writer — Type.

BED.Writer(output::IO)

Create a data writer of the BED file format.

Arguments:

output: data sink

source

BED.Record — Type.

BED.Record()

Create an unfilled BED record.

source

BED.Record(data::Vector{UInt8})

Create a BED record object from data.

This function verifies and indexes fields for accessors. Note that the ownership of data is transferred to a new record object.

source

BED.Record(str::AbstractString)

Create a BED record object from str.

This function verifies and indexes fields for accessors.

source

BED.chrom — Function.

chrom(record::Record)::String

Get the chromosome name of record.

source

BED.chromstart — Function.

chromstart(record::Record)::Int

Get the starting position of record.

Note that the first base is numbered 1.

source

BED.chromend — Function.

chromend(record::Record)::Int

Get the end position of record.

source

BED.name — Function.

name(record::Record)::String

Get the name of record.

source

BED.score — Function.

score(record::Record)::Int

Get the score between 0 and 1000.

source

BED.strand — Function.

strand(record::Record)::GenomicFeatures.Strand

Get the strand of record.

source

BED.thickstart — Function.

thickstart(record::Record)::Int

Get the starting position at which record is drawn thickly.

Note that the first base is numbered 1.

source

BED.thickend — Function.

thickend(record::Record)::Int

Get the end position at which record is drawn thickly.

source

BED.itemrgb — Function.

itemrgb(record::Record)::ColorTypes.RGB

Get the RGB value of record.

The return type is defined in ColorTypes.jl.

source

BED.blockcount — Function.

blockcount(record::Record)::Int

Get the number of blocks (exons) in record.

source

BED.blocksizes — Function.

blocksizes(record::Record)::Vector{Int}

Get the block (exon) sizes of record.

source

BED.blockstarts — Function.

blockstarts(record::Record)::Vector{Int}

Get the block (exon) starts of record.

Note that the first base is numbered 1.

source