API Reference

Public

Internal

BED.Reader — Method

BED.Reader(input::IO; index=nothing)
BED.Reader(input::AbstractString; index=:auto)

Create a data reader of the BED file format.

The first argument specifies the data source. When it is a filepath that ends with .bgz, it is considered to be block compression file format (BGZF) and the function will try to find a tabix index file (<filename>.tbi) and read it if any. See http://www.htslib.org/doc/tabix.html for bgzip and tabix tools.

Arguments

input: data source

index: path to a tabix file

BED.Record — Method

BED.Record(str::AbstractString)

Create a BED record object from str.

This function verifies and indexes fields for accessors.

BED.Record — Method

BED.Record(data::Vector{UInt8})

Create a BED record object from data.

This function verifies and indexes fields for accessors. Note that the ownership of data is transferred to a new record object.

BED.Record — Method

BED.Record()

Create an unfilled BED record.

BED.Writer — Type

BED.Writer(output::IO)

Create a data writer of the BED file format.

Arguments:

output: data sink

BED.blockcount — Method

blockcount(record::Record)::Int

Get the number of blocks (exons) in record.

BED.blocksizes — Method

blocksizes(record::Record)::Vector{Int}

Get the block (exon) sizes of record.

BED.blockstarts — Method

blockstarts(record::Record)::Vector{Int}

Get the block (exon) starts of record.

Note that the first base is numbered 1.

BED.chrom — Method

chrom(record::Record)::String

Get the chromosome name of record.

BED.chromend — Method

chromend(record::Record)::Int

Get the end position of record.

BED.chromstart — Method

chromstart(record::Record)::Int

Get the starting position of record.

Note that the first base is numbered 1.

BED.itemrgb — Method

itemrgb(record::Record)::ColorTypes.RGB

Get the RGB value of record.

The return type is defined in ColorTypes.jl.

BED.name — Method

name(record::Record)::String

Get the name of record.

BED.score — Method

score(record::Record)::Int

Get the score between 0 and 1000.

BED.strand — Method

strand(record::Record)::GenomicFeatures.Strand

Get the strand of record.

BED.thickend — Method

thickend(record::Record)::Int

Get the end position at which record is drawn thickly.

BED.thickstart — Method

thickstart(record::Record)::Int

Get the starting position at which record is drawn thickly.

Note that the first base is numbered 1.